Changed controllers => added dependent models in url for each controller

+++ src/main/java/com/eEducation/ftn/web/controller/ColloquiumController.java
@@ -19,7 +19,7 @@ import com.eEducation.ftn.service.CourseService;
 import com.eEducation.ftn.web.dto.ColloquiumDTO;
 
 @RestController
-@RequestMapping(value="api/colloquiums")
+@RequestMapping(value="api/course/{courseId}/colloquiums")
 public class ColloquiumController {
 	@Autowired
 	ColloquiumService colloquiumService;
@@ -28,7 +28,12 @@ public class ColloquiumController {
 	CourseService courseService;
 	
 	@RequestMapping(method = RequestMethod.GET)
-	public ResponseEntity<List<ColloquiumDTO>> getAll(){
+	public ResponseEntity<List<ColloquiumDTO>> getAll(@PathVariable Integer courseId){
+		Course course = courseService.findOne(courseId);
+		if(course == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+		
 		List<Colloquium> colloquiums = colloquiumService.findAll();
 		List<ColloquiumDTO> colloquiumDTOs = new ArrayList<>();
 		
@@ -40,7 +45,12 @@ public class ColloquiumController {
 	}
 	
 	@RequestMapping(method = RequestMethod.GET, value="/{id}")
-	public ResponseEntity<ColloquiumDTO> getById(@PathVariable Integer id){
+	public ResponseEntity<ColloquiumDTO> getById(@PathVariable Integer courseId, @PathVariable Integer id){
+		Course course = courseService.findOne(courseId);
+		if(course == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+		
 		Colloquium found = colloquiumService.findOne(id);
 		if(found == null){
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -50,15 +60,15 @@ public class ColloquiumController {
 	}
 	
 	@RequestMapping(method=RequestMethod.POST, consumes="application/json")
-	public ResponseEntity<ColloquiumDTO> save(@RequestBody ColloquiumDTO colloquium){
-		Colloquium newColloquium = new Colloquium();
-		
-		if(colloquium.getCourse() == null) {
+	public ResponseEntity<ColloquiumDTO> save(@RequestBody ColloquiumDTO colloquium, @PathVariable Integer courseId){
+		Course course = courseService.findOne(courseId);
+		if(course == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 		
-		Course course = courseService.findOne(colloquium.getCourse().getId());
-		if(course == null) {
+		Colloquium newColloquium = new Colloquium();
+		
+		if(colloquium.getCourse() == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 		
@@ -72,7 +82,12 @@ public class ColloquiumController {
 	}
 	
 	@RequestMapping(method=RequestMethod.PUT, consumes="application/json")
-	public ResponseEntity<ColloquiumDTO> update(@RequestBody ColloquiumDTO colloquium){
+	public ResponseEntity<ColloquiumDTO> update(@RequestBody ColloquiumDTO colloquium, @PathVariable Integer courseId){
+		Course course = courseService.findOne(courseId);
+		if(course == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+		
 		Colloquium found = colloquiumService.findOne(colloquium.getId());
 		if(found == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -89,7 +104,12 @@ public class ColloquiumController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Void> delete(@PathVariable Integer id){
+	public ResponseEntity<Void> delete(@PathVariable Integer id, @PathVariable Integer courseId){
+		Course course = courseService.findOne(courseId);
+		if(course == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+		
 		Colloquium found = colloquiumService.findOne(id);
 		if(found != null){
 			colloquiumService.remove(id);
